requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
pydantic==2.7.4
python-multipart==0.0.9
